requirements.txt
beautifulsoup4==4.12.3
boilerpy3==1.0.7
Requests==2.31.0
openai==1.23.2
python-dotenv==1.0.1